move cfg logging to logging.py and fix bugs for template_script
